+++ src/gspits/noninteracting/__init__.py
@@ -1,0 +1,1 @@
+"""Routines focused on the physical properties of non-interacting gases."""